+++ database/migrations/2026_05_05_111630_create_comment_table.php
@@ -13,9 +13,9 @@ return new class extends Migration
     {
         Schema::create('comment', function (Blueprint $table) {
             $table->id();
-            $table->integer('post_id');
+            $table->unsignedBigInteger('post_id');
             $table->text('content');
-            $table->integer('created_by');
+            $table->unsignedBigInteger('created_by');
             $table->foreign('post_id')->references('id')->on('post')->onDelete('cascade');
             $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
